refactor: split uml_state and ast_cache index orchestration
uml_state.py:
- Thin public wrapper only (~30 lines)
- Delegates to _uml_state_helpers.py for all implementation
- build_state_result accepts injectable parse_fn for test monkeypatching

_uml_state_helpers.py (new):
- Full UML state scanner implementation moved from uml_state.py
- Resolves health hot spot: was 421 lines in a single file

ast_cache.py:
- index_project() now delegates to cache/indexer.run_index_project()
- Removed ~235 lines of index orchestration from god-file (989 to 764 lines)
- Preserves back-compat module-level exports for tests/file_watcher/incremental_sync

cache/indexer.py:
- Added run_index_project, index_parallel, post_index_backfill, _walk_source_files
- Owns all project indexing orchestration (+204 lines)
- Runtime import of ast_cache_mod for monkeypatch-compatible delegate calls
